+++ drive/hooks.py
@@ -19,7 +19,6 @@ add_to_apps_screen = [
         "logo": "/assets/drive/frontend/favicon-310x310.png",
         "title": "Drive",
         "route": "/drive",
-        "has_permission": "drive.api.permissions.has_app_permission",
     }
 ]
