+++ frontend/src/components/FileUploader.vue
@@ -6,7 +6,6 @@ import { ref, onMounted, onBeforeUnmount, inject, watch } from "vue"
 import { useStore } from "vuex"
 import { useRoute } from "vue-router"
 import Dropzone from "dropzone"
-import { capture } from "@/telemetry"
 
 const store = useStore()
 const route = useRoute()
