+++ frontend/src/components/PrimaryDropdown.vue
@@ -76,7 +76,7 @@ import Users from "@/components/EspressoIcons/Users.vue"
 import AppSwitcher from "@/components/AppSwitcher.vue"
 import TeamSwitcher from "@/components/TeamSwitcher.vue"
 import { getTeams } from "@/resources/files"
-import emitter from "@/event-bus"
+import emitter from "@/emitter"
 import { ref, computed } from "vue"
 import { useStore } from "vuex"
 import { useRouter, useRoute } from "vue-router"
