--- frontend/test_package.json
@@ -1,91 +1,0 @@
-{
-  "name": "frontend",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "debug": "vite build --mode=development --base=/assets/drive/frontend/ && yarn copy-html-entry",
-    "build": "vite build --base=/assets/drive/frontend/ && yarn copy-html-entry",
-    "copy-html-entry": "cp ../drive/public/frontend/index.html ../drive/www/drive.html",
-    "preview": "vite preview",
-    "precommit": "pretty-quick --staged",
-    "prepare": "cd .. && husky install frontend/.husky",
-    "lint": "eslint --fix --ext .js,.vue ."
-  },
-  "dependencies": {
-    "@fontsource/geist-mono": "^5.0.2",
-    "@fontsource/ibm-plex-mono": "^5.0.12",
-    "@fontsource/lora": "^5.0.16",
-    "@fontsource/nunito": "^5.0.17",
-    "@headlessui-float/vue": "^0.13.3",
-    "@headlessui/vue": "^1.7.8",
-    "@hocuspocus/transformer": "^2.13.5",
-    "@tato30/vue-pdf": "^1.10.0",
-    "@tiptap/core": "2.6.6",
-    "@tiptap/extension-character-count": "2.6.6",
-    "@tiptap/extension-collaboration": "2.6.6",
-    "@tiptap/extension-collaboration-cursor": "2.6.6",
-    "@tiptap/extension-font-family": "2.6.6",
-    "@tiptap/extension-hard-break": "2.6.6",
-    "@tiptap/extension-link": "2.6.6",
-    "@tiptap/extension-placeholder": "2.6.6",
-    "@tiptap/extension-table": "2.6.6",
-    "@tiptap/extension-table-header": "2.6.6",
-    "@tiptap/extension-table-row": "2.6.6",
-    "@tiptap/extension-task-item": "2.6.6",
-    "@tiptap/extension-task-list": "2.6.6",
-    "@tiptap/extension-text-align": "2.6.6",
-    "@tiptap/extension-text-style": "2.6.6",
-    "@tiptap/extension-typography": "2.6.6",
-    "@tiptap/extension-underline": "2.6.6",
-    "@tiptap/pm": "2.6.6",
-    "@tiptap/starter-kit": "2.6.6",
-    "@tiptap/suggestion": "2.6.6",
-    "@tiptap/vue-3": "2.6.6",
-    "@uppy/core": "^3.9.3",
-    "@uppy/drop-target": "^2.0.4",
-    "@uppy/tus": "^3.5.3",
-    "@vueuse/core": "^10.3.0",
-    "canvas-datagrid": "^0.4.5",
-    "date-fns": "^2.30.0",
-    "docx-preview": "^0.1.15",
-    "dropzone": "^6.0.0-beta.2",
-    "feather-icons": "^4.29.0",
-    "frappe-ui": "^0.1.62",
-    "fuzzysort": "^2.0.4",
-    "idb-keyval": "^6.2.1",
-    "js-base64": "^3.7.5",
-    "jszip": "^3.10.1",
-    "lucide-vue-next": "^0.336.0",
-    "mammoth": "^1.6.0",
-    "mitt": "^3.0.0",
-    "sass": "^1.62.0",
-    "uuid": "^9.0.0",
-    "vue": "^3.2",
-    "vue-inline-svg": "^2.1.3",
-    "vue-router": "^4.1.6",
-    "vue-tippy": "^6.4.1",
-    "vuex": "^4.1.0",
-    "x-data-spreadsheet": "^1.1.9",
-    "xlsx": "^0.18.5",
-    "y-indexeddb": "^9.0.11",
-    "y-prosemirror": "^1.2.12",
-    "y-webrtc": "^10.3.0",
-    "yjs": "^13.6.18"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^4.0.0",
-    "autoprefixer": "^10.4.13",
-    "eslint": "^8.35.0",
-    "eslint-config-prettier": "^8.7.0",
-    "eslint-plugin-vue": "^9.9.0",
-    "husky": "^8.0.3",
-    "less": "^4.2.0",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "pretty-quick": "^3.1.3",
-    "tailwindcss": "^3.2.4",
-    "vite": "^4.0",
-    "vite-plugin-top-level-await": "^1.4.1"
-  }
-}
