--- package.json
@@ -1,8 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "postinstall": "cd frontend && npm install",
-    "dev": "cd frontend && npm run dev",
-    "build": "cd frontend && npm run build"
-  }
-}
